modularise components and improved responsiveness
