+++ src/client/assets/landing-page-graphics/landing-main.png
